+++ src/aux-common/Files/FileCalculations.spec.ts
@@ -64,6 +64,7 @@ import {
     isContextLocked,
     isDestroyable,
     isEditable,
+    normalizeAUXFileURL,
 } from './FileCalculations';
 import { cloneDeep } from 'lodash';
 import { File, Object, PartialFile } from './File';
@@ -1731,7 +1732,7 @@ describe('FileCalculations', () => {
         });
     });
 
-    describe('createWorkspace', () => {
+    describe('createWorkspace()', () => {
         it('should create new random context id if empty', () => {
             uuidMock.mockReturnValue('uuid');
             const workspace = createWorkspace('test', '');
@@ -1766,6 +1767,13 @@ describe('FileCalculations', () => {
             const workspace = createWorkspace('test', 'userSetID');
 
             expect(workspace.tags['aux.context.locked']).toEqual(true);
+        });
+
+        it('should allow setting the workspace to be unlocked', () => {
+            uuidMock.mockReturnValue('uuid');
+            const workspace = createWorkspace('test', 'userSetID', false);
+
+            expect(workspace.tags['aux.context.locked']).toEqual(false);
         });
     });
 
@@ -2898,6 +2906,28 @@ describe('FileCalculations', () => {
         });
     });
 
+    describe('normalizeAUXFileURL()', () => {
+        const cases = [
+            ['http://example.com/path', 'http://example.com/path.aux'],
+            ['http://example.com/', 'http://example.com/.aux'],
+            ['http://example.com', 'http://example.com/.aux'],
+            ['https://example.com/*/test', 'https://example.com/*/test.aux'],
+            [
+                'http://example.com/context/channel',
+                'http://example.com/context/channel.aux',
+            ],
+            [
+                'http://example.com/context/channel.aux',
+                'http://example.com/context/channel.aux',
+            ],
+            ['http://example.com/.aux', 'http://example.com/.aux'],
+        ];
+
+        it.each(cases)('should map %s to %s', (given, expected) => {
+            expect(normalizeAUXFileURL(given)).toBe(expected);
+        });
+    });
+
     describe('validateTag()', () => {
         it('should return invalid when tag is empty or null', () => {
             let errors = validateTag('');
@@ -3217,6 +3247,31 @@ describe('FileCalculations', () => {
                 '_hiddenTag4',
                 'other',
             ]);
+        });
+    });
+
+    describe('getFileShape()', () => {
+        const cases = [['cube'], ['sphere'], ['sprite']];
+        it.each(cases)('should return %s', (shape: string) => {
+            const file = createFile('test', {
+                'aux.shape': <any>shape,
+            });
+
+            const calc = createCalculationContext([file]);
+
+            expect(getFileShape(calc, file)).toBe(shape);
+        });
+
+        it('should return sphere when the file is a diff', () => {
+            const file = createFile('test', {
+                'aux.shape': 'cube',
+                'aux._diff': true,
+                'aux._diffTags': ['aux.shape'],
+            });
+
+            const calc = createCalculationContext([file]);
+
+            expect(getFileShape(calc, file)).toBe('sphere');
         });
     });