+++ src/aux-server/aux-web/aux-projector/FileTable/FileTable.ts
@@ -20,6 +20,8 @@ import {
     getAllFileTags,
     toast,
     isEditable,
+    createContextId,
+    addToContextDiff,
 } from '@casual-simulation/aux-common';
 import { EventBus } from '../../shared/EventBus';
 import { appManager } from '../../shared/AppManager';
@@ -34,6 +36,7 @@ import { TreeView } from 'vue-json-tree-view';
 import { downloadAuxState } from '../download';
 import { storedTree, site } from '@casual-simulation/causal-trees';
 import Cube from '../public/icons/Cube.svg';
+import Hexagon from '../public/icons/Hexagon.svg';
 
 @Component({
     components: {
@@ -44,6 +47,7 @@ import Cube from '../public/icons/Cube.svg';
         'file-table-toggle': FileTableToggle,
         'tree-view': TreeView,
         'cube-icon': Cube,
+        'hex-icon': Hexagon,
     },
 })
 export default class FileTable extends Vue {
@@ -84,6 +88,10 @@ export default class FileTable extends Vue {
     blacklistIndex: boolean[] = [];
     blacklistCount: number[] = [];
     editableMap: Map<string, boolean>;
+
+    showCreateWorksurfaceDialog: boolean = false;
+    worksurfaceContext: string = '';
+    worksurfaceAllowPlayer: boolean = true;
 
     uiHtmlElements(): HTMLElement[] {
         if (this.$refs.tags) {
@@ -290,6 +298,58 @@ export default class FileTable extends Vue {
         }
     }
 
+    public createSurface(): void {
+        this.worksurfaceContext = createContextId();
+        this.worksurfaceAllowPlayer = true;
+        this.showCreateWorksurfaceDialog = true;
+    }
+
+    /**
+     * Confirm event from the create worksurface dialog.
+     */
+    async onConfirmCreateWorksurface() {
+        this.showCreateWorksurfaceDialog = false;
+        const workspace = await this.fileManager.helper.createWorkspace(
+            undefined,
+            this.worksurfaceContext,
+            !this.worksurfaceAllowPlayer
+        );
+
+        if (!this.diffSelected) {
+            const calc = this.fileManager.helper.createContext();
+            for (let i = 0; i < this.files.length; i++) {
+                const file = this.files[i];
+                await this.fileManager.helper.updateFile(file, {
+                    tags: {
+                        ...addToContextDiff(
+                            calc,
+                            this.worksurfaceContext,
+                            0,
+                            0,
+                            i
+                        ),
+                    },
+                });
+            }
+        }
+
+        await this.fileManager.selection.selectFile(workspace, true);
+
+        this.resetCreateWorksurfaceDialog();
+    }
+
+    /**
+     * Cancel event from the create worksurface dialog.
+     */
+    onCancelCreateWorksurface() {
+        this.resetCreateWorksurfaceDialog();
+    }
+
+    resetCreateWorksurfaceDialog() {
+        this.showCreateWorksurfaceDialog = false;
+        this.worksurfaceAllowPlayer = true;
+    }
+
     onTagChanged(file: AuxObject, tag: string, value: string) {
         this.lastEditedTag = this.focusedTag = tag;
         this.focusedFile = file;
@@ -363,7 +423,6 @@ export default class FileTable extends Vue {
 
     async clearDiff() {
         await this.fileManager.recent.clear();
-        this.fileManager.recent.selectedRecentFile = this.fileManager.recent.files[0];
     }
 
     constructor() {